[A2] Task 2 & 3 change #ifdef sections
Now uses `TASK1` for task 2 and `TASK2` for task 3. This is because in
the file, task 2 is referred to as task 1 and task 3 as task 2.

Also change ways the macros are defined when calling make. Instead of
defaulting to being on, now needs to be explicitly passed either as
argument to `make` or exported as env var. The value of the varaible
does not matter.